Move AstronomicalConstant to it's own file. Tests as well.
